requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
djangorestframework==3.14.0
drf-yasg==1.21.7
Faker==19.11.0
flake8==6.1.0
idna==3.4
inflection==0.5.1
mccabe==0.7.0
packaging==23.2
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
